Splitting the parallel VM operations scenarios into two samples
